chore: move ui logic into ui folder
