--- gameplay/game/flaggedNames.js
@@ -1,3 +1,0 @@
-const flaggedNames = ["ASS", "FAG", "SUX", "DIE", "SEX", "PMS", "WTF", "CUM", "JAP", "DUM"];
-
-export default flaggedNames
