+++ gameplay/ui/flaggedNames.js
@@ -1,0 +1,27 @@
+export const flaggedNames = [
+    "ASS", "AZZ",
+    
+    "BUM",
+
+    "COK", "CUM", 
+    
+    "DUM", "DIE", "DIK", "DYK",
+    
+    "FAG", 
+    
+    "JAP",     
+    
+    "KKK", "KOK", 
+    
+    "NIG",
+
+    "PIS",
+    "POO",
+
+    "SEX", "SUK", "SUX", 
+    
+    "TIT", 
+    
+    "WTF", 
+    
+    "XXX"];
